+++ CHANGELOG.md
@@ -182,6 +182,9 @@ Adding a new version? You'll need three changes:
 - Admin and proxy listens in the deploy manifests now use the same parameters
   as the default upstream kong.conf.
   [#3165](https://github.com/Kong/kubernetes-ingress-controller/pull/3165)
+- Fix the behavior of filtering hostnames in `HTTPRoutes` when listeners 
+  of parent gateways specified hostname.
+  [#3180](https://github.com/Kong/kubernetes-ingress-controller/pull/3180)
 
 ## [2.7.0]